feature: parse escape sequences in Strings
This parses and tests escape sequences like '\n' in JS strings. It's
definetly not fulfilling the standard, but that's not a rabbit hole I'm
willing to go down and it is enough for my use case atm.

Thanks to this beautiful SO question, that helped me write it:
https://codereview.stackexchange.com/questions/2406/parsing-strings-with-escaped-characters-using-parsec

add rudimentary code for parsing Unicode escape sequences